feat(#69) :: 챌린지 참여 API 제작
